// src/Interfaces/SAPDFormParams.ts
export interface SAPDFormParams{
	ToolkitScriptManager1_HiddenField: string;
	__EVENTTARGET: string;
	__EVENTARGUMENT: string;
	__LASTFOCUS: string;
	__VIEWSTATE: string;
	__VIEWSTATEGENERATOR: string;
	__EVENTVALIDATION: string;
	txtStart: string;
	rdbSearchRange: string;
	txtEndDate: string;
	txtZipcode: string;
	ddlCategory: string;
	ddlCouncilDistrict: string;
	ddlSchoolDistrict: string;
	cbxHOA$cbxHOA_TextBox: string;
	cbxHOA$cbxHOA_HiddenField: string;
	btnSearch: string;
}